snpgdsCombineGeno {SNPRelate}	R Documentation

Merge SNP datasets

Description

To merge GDS files of SNP genotypes into a single GDS file

Usage

snpgdsCombineGeno(gds.fn, out.fn, sample.id=NULL, snpobj=NULL,
    name.prefix=NULL, snpfirstdim=TRUE, compress.annotation="ZIP_RA.MAX",
    compress.geno="", other.vars=NULL, verbose=TRUE)

Arguments

gds.fn
a list of SNP GDS files to be merged

out.fn
the name of output GDS file

sample.id
NULL, or a list. If it is a list, specify sample ids for each SNP GDS file

snpobj
specify a snpgdsSNPListClass object, used for strand switch; if NULL, the strand information of the first SNP GDS file is used

name.prefix
NULL, a character vector (added to sample ids for each GDS file)

snpfirstdim
if TRUE, genotypes are stored in the individual-major mode, (i.e, list all SNPs for the first individual, and then list all SNPs for the second individual, etc)

compress.annotation
the compression method for the variables except genotype

compress.geno
the compression method for the variable genotype

other.vars
a list object storing other variables

verbose
if TRUE, show information

Details

The typical variables specified in other.vars are “sample.annot” and “snp.annot”, which are data.frame objects.

WARNING: the functionality of snpgdsCombineGeno() is limited.

Value

None.

Author(s)

Xiuwen Zheng

See Also

snpgdsCreateGeno, snpgdsCreateGenoSet

Examples

# get the file name of a gds file
fn <- snpgdsExampleFileName()

# combine
snpgdsCombineGeno(c(fn, fn), "test.gds")

snpgdsSummary("test.gds")

[Package SNPRelate version 1.10.2 Index]