snpgdsIndInb {SNPRelate}	R Documentation

Individual Inbreeding Coefficients

Description

To calculate individual inbreeding coefficients using SNP genotype data

Usage

snpgdsIndInb(gdsobj, sample.id=NULL, snp.id=NULL,
    autosome.only=TRUE, remove.monosnp=TRUE, maf=NaN, missing.rate=NaN,
    method=c("mom.weir", "mom.visscher", "mle"), allele.freq=NULL,
    out.num.iter=TRUE, reltol=.Machine$double.eps^0.75, verbose=TRUE)

Arguments

gdsobj
an object of class SNPGDSFileClass, a SNP GDS file

sample.id
a vector of sample id specifying selected samples; if NULL, all samples are used

snp.id
a vector of snp id specifying selected SNPs; if NULL, all SNPs are used

autosome.only
if TRUE, use autosomal SNPs only; if it is a numeric or character value, keep SNPs according to the specified chromosome

remove.monosnp
if TRUE, remove monomorphic SNPs

maf
to use the SNPs with ">= maf" only; if NaN, no MAF threshold

missing.rate
to use the SNPs with "<= missing.rate" only; if NaN, no missing threshold

method
see details

allele.freq
to specify the allele frequencies; if NULL, the allele frequencies are estimated from the given samples

out.num.iter
output the numbers of iterations

reltol
relative convergence tolerance used in MLE; the algorithm stops if it is unable to reduce the value of log likelihood by a factor of $reltol * (abs(log likelihood with the initial parameters) + reltol)$ at a step.

verbose
if TRUE, show information

Details

The method can be: "mom.weir": a modified Visscher's estimator, proposed by Bruce Weir; "mom.visscher": Visscher's estimator described in Yang et al. (2010); "mle": the maximum likelihood estimation.

Value

Return estimated inbreeding coefficient.

Author(s)

Xiuwen Zheng

References

Yang J, Benyamin B, McEvoy BP, Gordon S, Henders AK, Nyholt DR, Madden PA, Heath AC, Martin NG, Montgomery GW, Goddard ME, Visscher PM. 2010. Common SNPs explain a large proportion of the heritability for human height. Nat Genet. 42(7):565-9. Epub 2010 Jun 20.

Examples

# open an example dataset (HapMap)
genofile <- snpgdsOpen(snpgdsExampleFileName())

rv <- snpgdsIndInb(genofile, method="mom.visscher")
head(rv$inbreeding)
summary(rv$inbreeding)

# close the genotype file
snpgdsClose(genofile)

[Package SNPRelate version 1.10.2 Index]