snpgdsPED2GDS {SNPRelate}	R Documentation

Conversion from PLINK PED to GDS

Description

Convert a PLINK PED text file to a GDS file.

Usage

snpgdsPED2GDS(ped.fn, map.fn, out.gdsfn, family=TRUE, snpfirstdim=FALSE,
    compress.annotation="ZIP_RA.max", compress.geno="", verbose=TRUE)

Arguments

ped.fn
the file name of PED file, genotype information

map.fn
the file name of MAP file

out.gdsfn
the output GDS file

family
if TRUE, to include family information in the sample annotation

snpfirstdim
if TRUE, genotypes are stored in the individual-major mode, (i.e, list all SNPs for the first individual, and then list all SNPs for the second individual, etc)

compress.annotation
the compression method for the GDS variables, except "genotype"; optional values are defined in the function add.gdsn

compress.geno
the compression method for "genotype"; optional values are defined in the function add.gdsn

verbose
if TRUE, show information

Details

GDS – Genomic Data Structures, the extended file name used for storing genetic data, and the file format is used in the gdsfmt package.

PED – PLINK PED format.

Value

None.

Author(s)

Xiuwen Zheng

References

Purcell S, Neale B, Todd-Brown K, Thomas L, Ferreira MAR, Bender D, Maller J, Sklar P, de Bakker PIW, Daly MJ & Sham PC. 2007. PLINK: a toolset for whole-genome association and population-based linkage analysis. American Journal of Human Genetics, 81.

See Also

snpgdsGDS2PED, snpgdsBED2GDS, snpgdsGDS2BED

Examples

# open
genofile <- snpgdsOpen(snpgdsExampleFileName())

snpgdsGDS2PED(genofile, "tmp")

# close
snpgdsClose(genofile)

# PED ==> GDS
snpgdsPED2GDS("tmp.ped", "tmp.map", "test.gds")

# delete the temporary file
unlink(c("tmp.ped", "tmp.map", "test.gds"), force=TRUE)

[Package SNPRelate version 1.3.5 Index]