snpgdsGRM {SNPRelate}	R Documentation

Genetic Relationship Matrix (GRM) for SNP genotype data

Description

Eigen-analysis on the Genetic Relationship Matrix (GRM) for SNP genotype data.

Usage

snpgdsGRM(gdsobj, sample.id=NULL, snp.id=NULL, autosome.only=TRUE,
    remove.monosnp=TRUE, maf=NaN, missing.rate=NaN, num.thread=1L,
    with.id=TRUE, verbose=TRUE)

Arguments

gdsobj
an object of class SNPGDSFileClass, a SNP GDS file

sample.id
a vector of sample id specifying selected samples; if NULL, all samples are used

snp.id
a vector of snp id specifying selected SNPs; if NULL, all SNPs are used

autosome.only
if TRUE, use autosomal SNPs only; if it is a numeric or character value, keep SNPs according to the specified chromosome

remove.monosnp
if TRUE, remove monomorphic SNPs

maf
to use the SNPs with ">= maf" only; if NaN, no MAF threshold

missing.rate
to use the SNPs with "<= missing.rate" only; if NaN, no missing threshold

num.thread
the number of (CPU) cores used; if NA, detect the number of cores automatically

with.id
if TRUE, the returned value with sample.id and sample.id

verbose
if TRUE, show information

Value

Return a list if with.id = TRUE:

sample.id
the sample ids used in the analysis

snp.id
the SNP ids used in the analysis

GRM
the genetic relationship matrix

If with.id = FALSE, return the genetic relationship matrix without sample and SNP IDs.

Author(s)

Xiuwen Zheng

References

Yang, J., Lee, S. H., Goddard, M. E. & Visscher, P. M. GCTA: a tool for genome-wide complex trait analysis. American journal of human genetics 88, 76-82 (2011).

See Also

snpgdsIndInb, snpgdsFst

Examples

# open an example dataset (HapMap)
genofile <- snpgdsOpen(snpgdsExampleFileName())

# Genetic relationship matrix
grm <- snpgdsGRM(genofile)

# close the file
snpgdsClose(genofile)

[Package SNPRelate version 1.2.0 Index]